[UPDATE] 스케줄러 정리, 리포지토리 이름 변경, 테이블 맵핑관계 에러때문에 잠시 주석

+++ tft/src/main/java/com/grassparty/tft/Model/DAO/StatDeckDAO.java
@@ -17,7 +17,7 @@ public class StatDeckDAO {
     private boolean valid;
     // 외래키
     @Column
-    private int season_id;
+    private long season_id;
 
     // 챔피언 칼럼 1~80
     @Column
